ref: moved internal/pkg/core to internal/core
ref: moved pkg to /
