+++ internal/core/domain/file.go
