+++ internal/core/port/processor.go
@@ -1,6 +1,6 @@
 package port
 
-import "github.com/andreychano/compressor-golang/internal/pkg/core/domain"
+import "github.com/andreychano/compressor-golang/internal/core/domain"
 
 // Processor defines a contract for any compression algorithm
 type Processor interface {
